backend/foodgram/requirements.txt
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.3.0
pytz==2021.1
sqlparse==0.4.1
